refactor: extract shared timeline layout logic, document project timeline endpoint
- Move getMonthlyPixelPosition, getDateFromMonthlyPixelPosition,
  computeBarPositionMonthly, computeBarPositionWeekly,
  getResizeHandleVisibility and assignRow into src/lib/timelineLayout.ts,
  imported by both TimelineRow and ProjectTimelineRow (pure refactor,
  behavior unchanged)
- Move duplicated generateWeeks from useTimeline/useProjectTimeline into
  the same shared module
- Document GET /api/projects/timeline in docs/api-reference.md
